\name{spiders}
\alias{spiders}
\docType{data}
\title{
  Bavarian Forest Spider Data
}
\description{
  Abundance of six spider species on 190 plots in the Bavarian Forest
  National Park
}
\usage{data("spiders")}
\format{
  A data frame with 190 observations on the following 9 variables.
  \describe{
    \item{\code{Plot}}{experimental plot id}
    \item{\code{Elevation}}{elevation of the plot}
    \item{\code{Canopy_openess}}{openess of canopy at plot}
    \item{\code{Pardosa_ferruginea}}{number of Pardosa ferruginea observed}
    \item{\code{Harpactea_lepida}}{number of Harpactea lepida observed}
    \item{\code{Callobius_claustrarius}}{number of Callobius claustrarius observed}
    \item{\code{Coelotes_terrestris}}{number of Coelotes terrestris observed}
    \item{\code{Pardosa_lugubris}}{number of Pardosa lugubris observed}
    \item{\code{Pardosa_riparia}}{number of Pardosa riparia observed}
  }
}
\details{

To untangle effects of dead-wood addition and canopy openness on
non-saproxylic epigeal arthropods, Seibold et al (2016) exposed different 
amounts of logs and branches on 190 0.1-ha plots located in sunny or shady 
mixed montane forests and sampled epigeal arthropods over three years. The
data contain abundances of six spider species for all 190 plots, along with
plot elevation and canopy openess.

}
\source{

  Sebastian Seibold, Claus Bässler, Petr Baldrian, Lena Reinhard, Simon Thorn,
  Michael D. Ulyshen, Ingmar Weiß and Jörg Müller (2016). Dead-wood addition 
  promotes non-saproxylic epigeal arthropods but effects are mediated by canopy
  openness. \emph{Biological Conservation}, \bold{204}, 181--188,
  \doi{10.1016/j.biocon.2016.09.031}.

}
\keyword{datasets}
